django/requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
importlib_metadata==7.1.0
Markdown==3.6
mysqlclient==2.2.4
packaging==24.0
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.11.0
zipp==3.18.1
